+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K4516RCBXR78GHDVHAYXF2AJ"
+        "01K457JMT6FDZMX002B1VDHW3C"
     )
     r"""The id of the resource"""