+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -149,5 +149,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-02T10:57:37.591Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-02T12:48:58.948Z")
     r"""The date and time the resource was last updated"""