+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K4516RC9G6K1BK737MSCAFYP"
+        "01K457JMT4GKFE3TG8PEQPH5AH"
     )
     r"""The id of the resource"""